continuing restructure
