+++ vc/duplicate.py
