+++ vc/pseudo_generation.py
